Bugfix: infinite loop in tick-mark computation
For very deep zoom states, the increments between the steps cannot
be resolved by the precision of the used Double type.
This commit skips computation of the tick-marks in this case.
This prevents an infinite loop which additionally in some cases
allocated an infinite number tickMark objects, rendering the applicaton
unusable (see #26).
The missing tick marks are not really a problem, since in this zoom
level usually no useful data can be shown, but the user is still able
to navigate back to a useful zoom state.
At even deeper zomm levels there was also the problem of dataRange
getting virtually zero, which led to problems at a lot of places, so
the zoomer now refuses to zoom that deep.