Change Toolbar to not resize the plotting area
Changes Chart to extend HiddenSidesPane instead of our custom
SidePane. This way the toolbar overlays the chart instead of
resizing it each time the toolbar is shown.

Also changes the pinned indication to be a change in background
color instead of a Border.

Fixes a bug in ParameterMeasurement, where the dropdown menu
was only accessible when the toolbar was pinned.