Allow selecting DataSet for Parameter Measurement (#68)
* Allow selecting DataSet for Parameter Measurement

The DataSet selection used the list returned by chart.getAllDataSets().
This function always returns the same ObservableArrayList, but on each call,
the list is cleared and refilled. This happens all the time, because eg the axes
use it to rescale to the data in each render loop. When the list is cleared, the
selection model is reset.

This commit fixes this issue by initializing a private observableList with the one
returned by getAllDataSets(). Now the listView only updates when the dialog is
closed and reopened, but this is imho not too bad.

* minor FX thread violation fix
* minor fix for default axis label position for CENTER_HOR and _VER