Fix NPE in TableViewer when updating columns
There seems to be a NPE in the openJFX TableView code, which crashes
the table view when the columns are updated frequently.
This commit works around the issue, by never removing columns but only
updating existing ones with new titles and setting the visibility.
To safeguard against cases when there are too many DataSets, the numnber
of displayed DataSets is limited at 100.