fix NPE in TableViewer + enable EditAxis for Plugins added after Plugin
* also fixes concurrent access violation on plugin removal

Changing the number of visible columns 100 times triggers a bug [1] in
OpenJFX, so we have to work around by never removing columns from the
table.

Never hide Columns, just reduce width, so the NPE is not triggered
Only update table when it is visible, reduces load, but the less code is
running, the less bugs can occur.
Refresh the selection model on change of elements, unfortunately this
resets the selction.

[1] https://bugs.openjdk.java.net/browse/JDK-8217953


* added early/return null safe-guard in addMouseEventHandlers()
* N.B. also removed useless qualifier 'MyPopOver.this'
* added unit-testing and changed 'private -> protected' to allow testing
* N.B. minor fixes: zoomDuration hasn't been propagated to PopUp