Replace recomputeLimits by addding to or invalidating range
Replaces use of recomputeLimits in dataSet setters, because they are
often hotspots and recomputing the limits is expensive.
Instead for add operations, DataRange.add can be used or the data is
invalidated and recomputed the next time DataRange.getMin/Max/Length is
called or at the beginning of the next render cycle by the chart.