Fix DefaultAxisDescription for tests
There where 2 cornercases for the axis description to handle:
- DataSetUtils uses raw axesDescriptions without a dataSet, so so we
cannot allways assume that DataSet.getAxesDescriptions().size() equals
DataSet.getDimensions(). This commit checks and returns if
AxesDescriptions are missing
- DimReductionDataSet subscribes to the events of the underlying data
set, but calls getMin() when minIndex is changed, which deadlocks.
Removing the listener befor and after changing the index solves the
deadlock.