Remove write lock from recomputeLimits
It should be possible to call recomputeLimits from within readLock
protected code, as it will allways return the same result when the
dataset is not modified.
Because it is also called in writeLock protected code, we cannot use a
read lock, but the caller has to care for proper locking.