Fix handling of Labels in CircularDoubleErrorDataSet
Fixes handling of labels and styles in the CircularDoubleErrorDataSet.
Before part of the implementation would use HashMap based implementation
from AbstractErrorDataSet and the other part used a
CircularBuffer<String> implementation.
Also in some parts of the code the labels feature was named tag and
nonexisting tags where changed to return a default value and not null.
Additionally added a simple unit test for the dataset.