Fix overlappingLabels bug for inverted axes
To detect clashing labels, the AbstractAxis implementation checks if the
first or last two labels are clashing and if that is the case, it sets
labelsOverlapping and uses the OverlapPolicy to circumvent the clashes.
When axes are inverted, the positions of the labels is inverted and the
code always returned that the labels where overlapping.
The fix also checks if the axes are inverted before comparing their
positions.