made ProfilerInfoBox public API and updated SimplePerformanceMeter
The ProfilerInfoBox is intended to be placed into the ToolBar and
provides a quick indication of FX pulse and actual scene frame rates,
process and system cpu loads, as well as JDK/JavaFx and (soon) Chart-fx
versioning information.

N.B. updates are in view of modularising Chart-fx (see also PR #160)
Interface to 'com.sun' package interface may become deprecated and/or
may cause inconveniences for users that have not yet fully moved to
Jigsaw-type modularisation. To be revisited for the 'x.2.y' release.