Improve constructors and documentation for Histogram
Improves documentation for Histogram, explaining overflow bins and bin
indexing.

Improve the constructors of Histogram to allow the resulting Histogram
to either have the outer bounds or the bin centers of the first and last
bin on the given min/max value. previously, the values where shifted
from the given range by one half of the bin size.

adresses #246