Fix ErrorDataSetRenderer for inverted x-Axis
For inverted x-Axis, the Renderer first checked for the points to
render and then swapped them and tried to account for the fact that the
points might not be exactly on the axis bounds by adding a fixed extra
point to each side, which should have been two, because the code which
determines which points to include for normal axes also adds one in the
other direction.
This commit replaces this by the cleaner approach of obtaining xMax at
the left bound of an inverted axis and vice versa.