Use user agent style sheet for default css
When adding the default css to the charts getStylesheets() list, it
effectively makes it impossible for users to overwrite values in
chart.css from their scene's stylesheet. By setting it as the chart's
user style sheet, the default chart.css has a lower priority than the
stylesheets added by the user.