Revert CI to using pull_request instead of pull_request_target
Using pull_request_target not only uses the ci.yml from the PR's base
but also checks out the PR base by default, so the changes are not
actually tested.
We could work around this and try to manually check out the PR branch,
but this complicates the CI, has potential security implications and the
only benefit is that external PRs also get snapshots in maven-central.
The ability to run coverity on external PR's is not impacted by this
change as it is a separate workflow.