Fix padded auto grow ranging
With padding enabled auto grow ranging was broken (see #465, thanks to
@ennerf for reporting) because the padding was applied to the auto
range stored in the axis. Because of this the padding was added each
time the axis range was updated with grow ranging enabled.

This commit changes Abstract axis to directly use the result of
`autoRange()` instead of overwriting the autoRange property with it.

In XYChart the autoGrow code had to be changed to use the previous
computed range instead of the displayed range (which would include the
padding, leading to pileup).

fixes #465