CrosshairIndicator: use correct bounds for positioning
The computation of the label position used the bounds of the complete
chart instead of only the plotting area, which lead to the label or
parts of it being occluded by the axes.
This commit fixes this by ussing the bounds of only the plot area to
compute the position of the label

Thanks to @kromar777 for reporting and proposing the solution.

fixes #523