ArrayCache: use custom Key type to circumvent interning
Because the JVM caches Integer objects, if Integer is useed as the key
of a weak HashMap, some entries will never be released. This commit
changes to use a ArrayCacheSizeKey type, which does not have interning.
This is prefered to creating unique Integer instances with `new` as this
constructor is deprecated.

fixes #484