Histogram: correctly handle reset()
The reset method would first reset the last dimension via the super
method, and then for vertical 2D datasets reset the x-axis description.
Because the y axis is also deleted and contains the binning information,
the histogram will always underflow.
This can be seen in the Histogram2DimSample which will underflow on the
yProjection after some time.