Axis: expose method to recompute transformation cache
If the axis gets changed multiple times during event handling, but the
cache is only updated during layout, axes transforms from the event
handlers use outdated transformations.

This has the effect that e.g. panning only applies the first mouse event
in each pulse and the panning is a lot slower than the actual mouse
movement.

This adds a method to the axes which can be called by plugins which use
and update the axis range.