Zoomer: base zoom calculations on pixels
Previously the zoomer would perform the scroll-zoom and panning
calculations in data coordinates. This is identical to pixel coordinates
for linear axes, but for e.g. log axes this leads to strange behaviour
since it can result in negative data values.

This changes the Pan and pan and scroll behaviour, the rectangle zoom
was naturally always using screen coordinates anyway, so this should
also be more consistent.