Chart: position plugins using Panes instead of Groups
It seems that groups are not well suited to positioning nodes on top of
the graph. Replacing the Groups for the Plugins by Panes fixes an issue
where the zoom rectangle would shift to the upper left corner of the
chart area whenever there was a relayout while a rectangle zoom was
ongoing.