Bump versions and CI
Bump the main java version from 11 to 17 and add a 21 build.

Bump javafx to version 16, as starting with 17 the memory leak tests
start to fail. This happens as soon as a plugin is present in the chart
and it seems to be related to the listener on the chart property.

Add update github actions to current versions.